+++ priv/repo/migrations/20190715074821_alter_ingest_remove_deprecated_relations.exs
@@ -1,0 +1,8 @@
+defmodule TdIe.Repo.Migrations.AlterIngestRemoveDeprecatedRelations do
+  use Ecto.Migration
+
+  def change do
+    alter(table(:ingest_versions), do: remove(:related_to))
+    alter(table(:ingests), do: remove(:parent_id))
+  end
+end